+++ CHANGELOG.md
@@ -5,10 +5,28 @@
 ### Added
 
 - `Tween` functions overloads taking `TweenParams`
+- `TweenParams` default values in constructor
+- `TweenParams` overrides `ToString()`
+- Unit tests
+    - Ease
+    - Pause
+    - Resume
+    - Stop
+    - Ping-pong
+    - Loop
+    - Generate, apply and destroy
+        - `Translation`
+        - `Rotation`
+        - `NonUniformScale`
+
+### Changed
+
+- `TweenParams` parameters order in constructor
+- `TweenStopSystem` schedules structural changes to `EndSimulationEntityCommandBufferSystem` (was `BeginSimulationEntityCommandBufferSystem`)
+- Systems use `World` (was `World.DefaultGameObjectInjectionWorld`)
 
 ### Fixed
 
-- Systems used `World.DefaultGameObjectInjectionWorld`
 
 ## [0.2.0] - 2020.10.23